requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.0
anyio==3.6.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cycler==0.11.0
einops==0.6.1
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gradio==3.30.0
gradio-client==0.2.4
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inplace-abn==1.1.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.2
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.21.6
orjson==3.8.12
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
pycodestyle==2.10.0
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.30.0
scikit-learn==1.0.2
scipy==1.7.3
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
threadpoolctl==3.1.0
timm==0.6.13
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing-extensions==4.5.0
uc-micro-py==1.0.2
urllib3==2.0.2
uvicorn==0.22.0
websockets==11.0.3
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.15.0
